+++ crates/bot-client/src/handler.rs
@@ -1,12 +1,14 @@
-use crate::commands;
-
-use std::error::Error;
+use crate::{SharedConfig, SharedHttp};
+use bot_error::Error;
 
 use log::{debug, error, info, trace};
 use serenity::async_trait;
-use serenity::builder::{CreateEmbed, CreateInteractionResponseFollowup};
+use serenity::builder::{
+	CreateEmbed, CreateInteractionResponse, CreateInteractionResponseFollowup,
+	CreateInteractionResponseMessage,
+};
 use serenity::model::{
-	application::{Command, Interaction},
+	application::{Command, CommandInteraction, Interaction},
 	colour::Colour,
 	gateway::Ready,
 };
@@ -16,14 +18,47 @@ use serenity::prelude::{Context, EventHandler};
 pub struct Handler;
 
 impl Handler {
-	async fn register_commands(&self, ctx: &Context) -> Result<(), Box<dyn Error>> {
-		let commands = commands::to_vec();
+	async fn register_commands(&self, ctx: &Context) -> Result<(), Error> {
+		let commands = bot_commands::to_vec();
 		let commands_len = commands.len();
 		for command in commands {
 			Command::create_global_command(&ctx.http, command).await?;
 		}
 
 		debug!("Registered {} commands", commands_len);
+		Ok(())
+	}
+
+	/// Dispatch our commands from a [`CommandInteraction`]
+	async fn dispatch_command(ctx: &Context, command: &CommandInteraction) -> Result<(), Error> {
+		let command_name = command.data.name.as_str();
+
+		// grab our configuration & http client from the aether
+		let (http, config) = {
+			let read = ctx.data.read().await;
+			let http = read
+				.get::<SharedHttp>()
+				.ok_or("Couldn't get shared HTTP client! WHY??????")?
+				.clone();
+			let config = read
+				.get::<SharedConfig>()
+				.ok_or("Couldn't get shared bot configuration!")?
+				.clone();
+			(http, config)
+		};
+
+		match command_name {
+			"ping" => bot_commands::ping::respond(ctx, command).await?,
+			"track" => bot_commands::track::respond(ctx, &http, &config, command).await?,
+			_ => {
+				let message = CreateInteractionResponseMessage::new().content(format!(
+					"It doesn't look like you can use `{command_name}`. Sorry :("
+				));
+				let response = CreateInteractionResponse::Message(message);
+				command.create_response(&ctx, response).await?;
+			}
+		};
+
 		Ok(())
 	}
 }
@@ -36,7 +71,7 @@ impl EventHandler for Handler {
 			let command_name = &command.data.name;
 			trace!("Received command: {}", command_name);
 
-			if let Err(why) = commands::dispatch(&ctx, &command).await {
+			if let Err(why) = Handler::dispatch_command(&ctx, &command).await {
 				error!(
 					"Ran into an error while dispatching command {}:\n{why:?}",
 					command_name
