+++ crates/bot-consts/src/lib.rs
