+++ crates/bot-error/src/lib.rs
@@ -1,0 +1,1 @@
+pub type Error = Box<dyn std::error::Error + Send + Sync>;
