+++ crates/bot-http/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 authors = ["seth <getchoo at tuta dot io>"]
-description = "The HTTP backend for nixpkgs-tracker-bot"
+description = "HTTP client for nixpkgs-tracker-bot"
 repository = "https://github.com/getchoo/nixpkgs-tracker-bot"
 
 publish = false
