+++ crates/bot-jobs/Cargo.toml
@@ -1,23 +1,20 @@
 [package]
-name = "bot-client"
+name = "bot-jobs"
 version = "0.1.0"
 edition = "2021"
 
 authors = ["seth <getchoo at tuta dot io>"]
-description = "Discord client for nixpkgs-tracker-bot"
+description = "Background jobs for nixpkgs-tracker-bot"
 repository = "https://github.com/getchoo/nixpkgs-tracker-bot"
 
 publish = false
 
 [dependencies]
-bot-commands = { workspace = true }
 bot-config = { workspace = true }
 bot-consts = { workspace = true }
 bot-error = { workspace = true }
-bot-http = { workspace = true }
-bot-jobs = { workspace = true }
+git2 = { workspace = true, features = ["https"] }
 log = { workspace = true }
-serenity = { workspace = true }
 tokio = { workspace = true }
 
 [lints.rust]
