+++ crates/bot-jobs/src/lib.rs
@@ -1,4 +1,5 @@
-use crate::{Config, Error};
+use bot_config::Config;
+use bot_error::Error;
 
 use std::time::Duration;
 
@@ -7,8 +8,13 @@ use log::error;
 mod repo;
 
 /// Run our jobs an initial time, then loop them on a separate thread
+///
+/// # Errors
+///
+/// Will return [`Err`] if any jobs fail
 pub fn dispatch(config: Config) -> Result<(), Error> {
 	repo::fetch_or_update_repository(&config.nixpkgs_path)?;
+
 	tokio::spawn(async move {
 		loop {
 			tokio::time::sleep(Duration::from_secs(repo::TTL_SECS)).await;
